chugging through some errors
Spring boot has a few gradiences with my code. first I spelled defer wrong in the application.yaml, then it didn't like that I had _ in my variable names of the entity but mainly it was upset that I used the wrong import for Id, spring boot couldn't see that I'd marked the entity ids because I'd used the springbootframework and not jakarta. Once I swapped that it moved onto the next error which was that I'd named the wrong variable when I MappedBy my one to many relationship. There's still more to fix but that's all for this moment I have time for.